add links to BookingBar
