+++ src/client/components/pages/booking/utilsBookingBar.js
@@ -1,0 +1,27 @@
+const STEPS = {
+  "/booking/flight": 0,
+  "/booking/seat": 1,
+  "/booking/luggage": 2
+  //   "/booking/": 3,
+  //   "/booking/": 4,
+  //   "/booking/": 5
+};
+
+export const getCurrentStep = history => {
+  return STEPS[history.location.pathname];
+};
+
+export const getLink = number => {
+  return Object.keys(STEPS).find(key => STEPS[key] === number);
+};
+
+export const getSteps = () => {
+  return [
+    "Flight selection",
+    "Seat selection",
+    "Luggage",
+    "Personal data",
+    "Payment",
+    "Confirmation"
+  ];
+};
